chore: update README with project overview and setup
